analyzer/requirements.txt
wxPython>=4.2.0
openai>=1.0.0
python-dotenv>=1.0.0
